Added IntroPage and imported primereact and moved files to src level out of component
