+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "classnames": "^2.2.6",
     "firebase": "^5.1.0",
     "primereact": "^1.6.2",
     "proptypes": "^1.1.0",
