--- src/Components/App.js
@@ -1,26 +1,0 @@
-import React, { Component } from 'react'
-import {Route, Link} from 'react-router-dom'
-import './App.css'
-
-class App extends Component {
-  constructor () {
-    super()
-    this.state = {
-      login: false
-    }
-  }
-  render () {
-    return (
-      <div className='App'>
-        <header className='App-header'>
-          <h1 className='App-title'>Welcome to React</h1>
-        </header>
-        <p className='App-intro'>
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
-    )
-  }
-}
-
-export default App
